Refactor VBO VAO Index buffer classes, assimp as submodule, create mesh and model abstraction

+++ CgEngineCore/src/CgEngineCore/Rendering/VertexArray.cpp
@@ -56,8 +56,8 @@ namespace CGEngine{
 	{
 		bind();
 		indexbuffer.bind();
-		m_indicies_count = indexbuffer.get_count();
-
+		m_indicies_count = static_cast<size_t>(indexbuffer.get_count());
+		
 	}
 
 	void VertexArray::bind() const
